Move feed_market.dart to the common directory.
